requirements.txt
numpy ==2.0.2
pandas ==2.3.3
matplotlib ==3.10.6
scikit-learn ==1.7.2
pillow == 11.3.0
tensorflow == 2.18.1
keras == 3.8.0


